+++ src/clients/alphaVantage/parsers/index.ts
@@ -1,3 +1,4 @@
-export * from './search';
-export * from './base';
-export * from './quote';
+export * from "./search";
+export * from "./base";
+export * from "./quote";
+export * from "./time_daily";