+++ src/schema/index.ts
@@ -1,5 +1,6 @@
-import { mergeTypeDefs } from '@graphql-tools/merge';
-import { quoteDefs } from './quote';
-import { searchDefs } from './search';
+import { mergeTypeDefs } from "@graphql-tools/merge";
+import { quoteDefs } from "./quote";
+import { searchDefs } from "./search";
+import { dailyDefs } from "./time_daily/schema";
 
-export const typeDefs = mergeTypeDefs([searchDefs, quoteDefs]);
+export const typeDefs = mergeTypeDefs([searchDefs, quoteDefs, dailyDefs]);