+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Feature: Add a query that fetches the latest 100 daily records for a stock. ([#6](https://github.com/josh-W42/stonks-graphwatch/pull/6))
 - Fix: Cleaned some file structure to maintain consistency. Added comments for readability. Changed GetQuoteGQL query arg from input to symbol to better represent input. ([#5](https://github.com/josh-W42/stonks-graphwatch/pull/5))
 - Feature: Added a quote query for getting data for a single stock by accepting the symbol as input. ([#4](https://github.com/josh-W42/stonks-graphwatch/pull/4))
 - Dev: Added additional error checking for responses from alpha vantage API ([#3](https://github.com/josh-W42/stonks-graphwatch/pull/3))